refactor(council)!: extract the quorum plumbing, and lower the size ratchet
Closes the 89-line residual the previous commit surfaced, by the path the gate
itself prescribes: lowering through extraction, never a raise.

`ai_council/quorum_wiring.ts` takes the 166 lines that were the quorum seam —
the setting reader, the min-present reader, the attendance-event emitter, the
post-run presence re-derivation and the k-of-n banner. Same layer as
`qualification_wiring.ts`, and the code this branch's change already reached
into (`_postRunQuorum` gained and then shed a probes channel here). The
discriminator that makes it a boundary rather than budget accounting: every
function moved is about a k-of-n verdict and its reporting, and none is about
argv, client construction, or payload writing — which is what `council_cli.ts`
is for. `Dict` is redeclared locally rather than imported, so the new module
does not couple back to the file it was separated from.

council_cli.ts 4,311 → 3,980; it leaves the three-worst list it used to head.
The ratchet total 19,725 → 19,647, i.e. 78 BELOW the old baseline, so the
baseline is lowered to match — which is what the gate asks for on a gain and
the opposite of the raise it calls a defect. The entry records both steps and
why comment-trimming reached only part of the first one.

Own-orphans from the move are removed in the same diff: `QuorumSetting`,
`absentReasonFromCliFailure`, `appendQuorumEvent`, `QuorumAbsence` and
`QuorumEventPhase` lost their last reference in council_cli.ts.

One process correction worth recording. `npx tsc --noEmit -p tsconfig.json`
does NOT cover `src/scripts/council_cli.ts` — that file belongs to
`tsconfig.scripts.json`, which only `npm run typecheck` runs. A dropped
`evaluateQuorum` import passed that check and was caught by the test suite
instead. `npm run typecheck` is green here.

Verified: 922 tests green, `npm run typecheck` clean, eslint clean,
`check_source_size_budget` exit 0 and its baseline test passing.
